Reorganized http-related structures into a signle file
